more tidyups and sanitizations
