+++ src/p2p/data_transform.rs
@@ -1,3 +1,4 @@
+///! https://docs.rs/libp2p-gossipsub/latest/libp2p_gossipsub/trait.DataTransform.html
 use libp2p::gossipsub::{DataTransform, Message, RawMessage, TopicHash};
 use std::io::{Error, ErrorKind};
 use std::time::{SystemTime, UNIX_EPOCH};
